requirements.txt
streamlit
langchain
langchain-huggingface
python-dotenv
requests
tldextract
